MAkign a nice one stop shop kernel extension module
