+++ setup.py
@@ -263,14 +263,17 @@ extensions = [
     create_extension(
         "synthesizer.extensions.kernel",
         [
-            "src/synthesizer/extensions/kernel.cpp",
+            "src/synthesizer/extensions/kernel_extensions/kernel_bindings.cpp",
+            "src/synthesizer/extensions/kernel_extensions/projected_kernel.cpp",
+            "src/synthesizer/extensions/kernel_extensions/truncated_kernel.cpp",
+            "src/synthesizer/extensions/kernel_extensions/overlap_kernel.cpp",
             "src/synthesizer/extensions/integration.cpp",
             "src/synthesizer/extensions/property_funcs.cpp",
             "src/synthesizer/extensions/numpy_init.cpp",
         ],
         compile_flags=compile_flags,
         links=link_args,
-        include_dirs=include_dirs,
+        include_dirs=include_dirs + ["src/synthesizer/extensions"],
     ),
     create_extension(
         "synthesizer.extensions.weights",
